Send user to verifyIdentityOnboarding instead of ohNo on coldstart.
Before we sent users directly to the ohNo(ekycRejected) and pendingVerification screens if the state was correct on cold start, to make things a bit more similiar we send them to the verifyIdentityOnboarding screens first, then send them to the correct screen after selecting Scan IC / Jumio.
Note: For kyc flow with multiple options, it would be a good idea to show some kind of status in progress / rejected on the screen where you select identity verification method.