Load context in RootCoordinator from GraphQL endpoint.
- Updated Apollo from 0.10.1 to 0.11.1 to be able to set Authorization header using HTTPNetworkTransportPreflightDelegate using a custom networkTransport for ApolloClient

Added extension helpers for existing models to be able to convert the GraphQL models to our existing models, this is only during first step of migrationg over to GraphQL so that we don't do too much at once.
Also note that the REST API and GraphQL schema does not always map when it comes to Optional / Required variables. You can see this in the extensions added for our models where we have to force unwrap. Ideally we update the GraphQL schema to reflect the Required variables we expect in the REST API.