Load products from GraphQL.
Load products from GraphQL instead of REST endpoint.

Note: We still don't have presentation property, so now you won't be able to buy anything and the products to buy is listed with "???".

We might want to combine or rename ProductModel and Product.
ProductModel is the model used for data returned by our server, while Product is used to populate the Apple Pay products.