Started refactoring and found a bug.
I started by dumbing down the MyInfoSummaryViewController which was
still making API calls on its own. I found that the region we update was
hard-coded once we finish with that step which is not what we want. I
also found we were incorrectly ignoring dynamic links when a pending
email was not present even though the email did not matter by the time
we are handling a dynamic link. This
corrects the issues.