Update to new GraphQL schema.
- Remove liveUnsupportedRegion test since we are using Enum from the server schema, this cause should not happen on the client.
- Use RegionCode enum from graphQL schema instead of string wherever country code is used.
- Renamed all fragments to use Fields as postfix to be consistent.
- Update queries and mutations according to new schema and move all graphql files into separate folder.

Note: We still don't handle errors in a good way (the errors inside a successful GraphQL response) as well as optional types from GraphQL (the type can be null if there is an error inside errors, though we should handle the case where there are no errors and the expected data is null). Since this is not handled properly yet, we use a couple of force unwraps, which should be removed.
Also, need to test that the new GraphQL endpoints actually work as the legacy APIs does.