Select the user's current country by default.
This required moving the location permission step. If it is in the
initial steps, it will be skipped and we'll never get location
permissions for users who cold-start mid sign-up.

The location controller doesn't need promises anymore, it just needs to
keep up with the current country the user is in during signup.