Phase 2!
This sets up the new tab order where the middle tab is now Coverage.
Havard is working on the much nicer looking UI that will be used here,
but this allows you to see the general idea. You can now go through
onboarding for singapore and norway separately. This selection of
countries is fairly arbitrary to show that onboarding can be summoned
for any region we expect the server to support.