Fix issue with onboarding ekycing for wrong region
Since there's special case shortcuts to ekyc for singapore and malaysia
for everyone in Norway, it resulted in a case where they would try to do
ekyc and wait forever to be approved. In the system we could see they
*were* approved, but for Norway (even when that's not a region they
could select). This fixes RegionOnboardingCoordinator to only take in
region to avoid such confusion later.